Mahjong testing
Testing classes with mahjong.
Goal will be to create a functional tabletop, a well of tiles, four
player hands, and create the shuffling and assigning of a basic hand.

AI and winning conditions are not expected any time soon.  I'd want to
learn how to create a graphical interface before I would put that much
effort in.